requirements.txt
tinydb
discord.py==0.14.3
aiohttp
websockets
chardet